make webdrivers non-mandatory

+++ bzt/modules/_selenium.py
@@ -305,6 +305,14 @@ class WebDriver(RequiredTool):
     def _get_latest_version_from_inet(self):
         pass
 
+    def _download_driver(self, **kwargs):
+        # don't fail if driver can't be downloaded
+        try:
+            return self._download(**kwargs)
+        except:
+            if self.mandatory:
+                raise
+
 
 class ChromeDriver(WebDriver):
     VERSION = "96.0.4664.45"
@@ -328,12 +336,13 @@ class ChromeDriver(WebDriver):
         if not os.path.exists(_dir):
             os.makedirs(_dir)
 
-        dist = self._download(use_link=True)
-        unzip(dist, _dir)
-        os.remove(dist)
+        dist = self._download_driver(use_link=True)
+        if dist:
+            unzip(dist, _dir)
+            os.remove(dist)
 
-        if not is_windows():
-            os.chmod(self.tool_path, 0o755)
+            if not is_windows():
+                os.chmod(self.tool_path, 0o755)
 
 
 class GeckoDriver(WebDriver):
@@ -362,14 +371,15 @@ class GeckoDriver(WebDriver):
         if not os.path.exists(_dir):
             os.makedirs(_dir)
 
-        dist = self._download(use_link=True)
-        if self.download_link.endswith('.zip'):
-            self.log.info("Unzipping %s to %s", dist, _dir)
-            unzip(dist, self.get_dir())
-        else:
-            self.log.info("Untaring %s to %s", dist, _dir)
-            untar(dist, _dir)
-        os.remove(dist)
+        dist = self._download_driver(use_link=True)
+        if dist:
+            if self.download_link.endswith('.zip'):
+                self.log.info("Unzipping %s to %s", dist, _dir)
+                unzip(dist, self.get_dir())
+            else:
+                self.log.info("Untaring %s to %s", dist, _dir)
+                untar(dist, _dir)
+            os.remove(dist)
 
-        if not is_windows():
-            os.chmod(self.tool_path, 0o755)
+            if not is_windows():
+                os.chmod(self.tool_path, 0o755)
